Functioning! I think. Have to go to tutor cs160, going to run more tests and clean up + writeup at PCC Sylvania from laptop. May not finish til late tonight.\n

+++ LinkedList.h
@@ -50,7 +50,7 @@ class LLL
 
 		bool is_empty(void);
 		size_t length(void);
-		T& at(int index);
+		T& at(size_t index);
 		void push_back(T& new_data);
 		T* lookup(T key); //must not use reference to accept rvalue arg
 		bool remove(T to_remove);
@@ -60,7 +60,7 @@ class LLL
 		Node<T>* tail;
 
 		size_t count_nodes(Node<T>* list);
-		T& find_at_index(Node<T>* list, int index);
+		T& find_at_index(Node<T>* list, size_t index);
 		void copy_all(Node<T>* src, Node<T>*& dest, Node<T>*& dest_tail, Node<T>* last_dest);
 		void remove_all(Node<T>* to_del);
 		void push_back(Node<T>* list, Node<T>* new_node);
@@ -79,8 +79,8 @@ class CLL
 		CLL(const CLL& src);
 		const CLL& operator=(const CLL& rhs);
 		bool is_empty(void);
-		int length(void);
-		T& at(int index);
+		size_t length(void);
+		T& at(size_t index);
 		void push_back(T& new_data);
 		T* lookup(T key);
 		bool remove(T to_remove);
@@ -89,8 +89,8 @@ class CLL
 		Node<T>* head;
 		Node<T>* tail;
 
-		int count_nodes(Node<T>* list);
-		T& find_at_index(Node<T>* list, int index);
+		size_t count_nodes(Node<T>* list);
+		T& find_at_index(Node<T>* list, size_t index);
 		void copy_all(Node<T>* src, Node<T>*& dest, Node<T>*& dest_tail,
 					  Node<T>* last_dest, Node<T>* src_head, Node<T>* dest_head);
 		void remove_all(Node<T>* to_delete);
